refactor: :truck: separate span factory and context files and tests
