+++ packages/core/lib/index.ts
@@ -12,6 +12,7 @@ export * from './processor'
 export * from './retry-queue'
 export * from './span'
 export * from './span-context'
+export * from './span-factory'
 export * from './time'
 export * from './validation'
 export { default as traceIdToSamplingRate } from './trace-id-to-sampling-rate'
