+++ packages/core/tests/span.test.ts
@@ -101,111 +101,6 @@ describe('SpanInternal', () => {
   })
 })
 
-describe('SpanFactory', () => {
-  describe('startSpan', () => {
-    it('omits first class span attribute by default', () => {
-      const clock = new IncrementingClock('1970-01-01T00:00:00.000Z')
-      const sampler = new Sampler(0.5)
-      const delivery = { send: jest.fn() }
-      const processor = { add: (span: SpanEnded) => delivery.send(spanToJson(span, clock)) }
-      const backgroundingListener = new ControllableBackgroundingListener()
-      const spanFactory = new SpanFactory(
-        processor,
-        sampler,
-        new StableIdGenerator(),
-        spanAttributesSource,
-        new IncrementingClock(),
-        backgroundingListener,
-        jestLogger,
-        new DefaultSpanContextStorage(backgroundingListener)
-      )
-
-      const span = spanFactory.startSpan('name')
-
-      // @ts-expect-error 'attributes' is private but very awkward to test otherwise
-      expect(span.attributes.attributes.has('bugsnag.span.first_class')).toBe(false)
-    })
-
-    it('creates first class spans when isFirstClass is true', () => {
-      const clock = new IncrementingClock('1970-01-01T00:00:00.000Z')
-      const sampler = new Sampler(0.5)
-      const delivery = { send: jest.fn() }
-      const processor = { add: (span: SpanEnded) => delivery.send(spanToJson(span, clock)) }
-      const backgroundingListener = new ControllableBackgroundingListener()
-      const spanFactory = new SpanFactory(
-        processor,
-        sampler,
-        new StableIdGenerator(),
-        spanAttributesSource,
-        new IncrementingClock(),
-        backgroundingListener,
-        jestLogger,
-        new DefaultSpanContextStorage(backgroundingListener)
-      )
-
-      const span = spanFactory.startSpan('name', { isFirstClass: true })
-
-      // @ts-expect-error 'attributes' is private but very awkward to test otherwise
-      expect(span.attributes.attributes.get('bugsnag.span.first_class')).toBe(true)
-    })
-
-    it('does not create first class spans when isFirstClass is false', () => {
-      const clock = new IncrementingClock('1970-01-01T00:00:00.000Z')
-      const sampler = new Sampler(0.5)
-      const delivery = { send: jest.fn() }
-      const processor = { add: (span: SpanEnded) => delivery.send(spanToJson(span, clock)) }
-      const backgroundingListener = new ControllableBackgroundingListener()
-      const spanFactory = new SpanFactory(
-        processor,
-        sampler,
-        new StableIdGenerator(),
-        spanAttributesSource,
-        new IncrementingClock(),
-        backgroundingListener,
-        jestLogger,
-        new DefaultSpanContextStorage(backgroundingListener)
-      )
-
-      const span = spanFactory.startSpan('name', { isFirstClass: false })
-
-      // @ts-expect-error 'attributes' is private but very awkward to test otherwise
-      expect(span.attributes.attributes.get('bugsnag.span.first_class')).toBe(false)
-    })
-
-    it.each([
-      null,
-      undefined,
-      1,
-      0,
-      'true',
-      'false',
-      [true, false]
-    ])('omits first class attribute when isFirstClass is %s', (isFirstClass) => {
-      const clock = new IncrementingClock('1970-01-01T00:00:00.000Z')
-      const sampler = new Sampler(0.5)
-      const delivery = { send: jest.fn() }
-      const processor = { add: (span: SpanEnded) => delivery.send(spanToJson(span, clock)) }
-      const backgroundingListener = new ControllableBackgroundingListener()
-      const spanFactory = new SpanFactory(
-        processor,
-        sampler,
-        new StableIdGenerator(),
-        spanAttributesSource,
-        new IncrementingClock(),
-        backgroundingListener,
-        jestLogger,
-        new DefaultSpanContextStorage(backgroundingListener)
-      )
-
-      // @ts-expect-error 'isFirstClass' is the wrong type
-      const span = spanFactory.startSpan('name', { isFirstClass })
-
-      // @ts-expect-error 'attributes' is private but very awkward to test otherwise
-      expect(span.attributes.attributes.has('bugsnag.span.first_class')).toBe(false)
-    })
-  })
-})
-
 describe('Span', () => {
   describe('client.startSpan()', () => {
     it('returns a Span', () => {
@@ -719,47 +614,6 @@ describe('SpanContext', () => {
 
       span.end()
       expect(span.isValid()).toEqual(false)
-    })
-  })
-  describe('spanContextEquals()', () => {
-    it.each([
-      {
-        span1: undefined,
-        span2: undefined,
-        expected: true
-      },
-      {
-        span1: { id: '0123456789abcdef', traceId: '0123456789abcdeffedcba9876543210', isValid: () => true },
-        span2: undefined,
-        expected: false
-      },
-      {
-        span1: undefined,
-        span2: { id: '0123456789abcdef', traceId: '0123456789abcdeffedcba9876543210', isValid: () => true },
-        expected: false
-      },
-      {
-        span1: { id: '0123456789abcdef', traceId: '0123456789abcdeffedcba9876543210', isValid: () => true },
-        span2: { id: '0123456789abcdef', traceId: '0123456789abcdeffedcba9876543210', isValid: () => true },
-        expected: true
-      },
-      {
-        span1: { id: '0123456789abcdef', traceId: '0123456789abcdeffedcba9876543210', isValid: () => true },
-        span2: { id: '0123456789abcdef', traceId: '0123456789abcdeffedcba9876543210', isValid: () => false },
-        expected: true
-      },
-      {
-        span1: { id: '0123456789abcdef', traceId: '0123456789abcdeffedcba9876543210', isValid: () => true },
-        span2: { id: '0123456789abcdef', traceId: 'a0b1c2d3e4f5a0b1c2d3e4f5a0b1c2d3', isValid: () => true },
-        expected: false
-      },
-      {
-        span1: { id: '0123456789abcdef', traceId: '0123456789abcdeffedcba9876543210', isValid: () => true },
-        span2: { id: '9876543210fedcba', traceId: '0123456789abcdeffedcba9876543210', isValid: () => true },
-        expected: false
-      }
-    ])('returns $expected given inputs $span1 and $span2', ({ span1, span2, expected }) => {
-      expect(spanContextEquals(span1, span2)).toEqual(expected)
     })
   })
 })
